requirements.txt
pandas
pytest
streamlit
spotipy
musicbrainzngs
google-generativeai
